Address PR feedback: add cert constant, move extensions to KeyVaultHelper, update docs
